:construction: stream functionality to mixin - TODO: rm host dep i.e., _mgnifier_helper
