+++ mgnipy/V2/datasets.py
@@ -1,33 +1,26 @@
 import asyncio
-import io
 import logging
 import warnings
-import webbrowser
-import zlib
 from pathlib import Path
 from typing import (
     Any,
-    Generator,
-    Literal,
     Optional,
 )
 
 import aiofiles
 import httpx
-import ijson
 import pandas as pd
 from pydantic import (
     DirectoryPath,
     HttpUrl,
 )
-from requests.exceptions import HTTPError
-from skbio.io import read
 from tqdm import tqdm as tqdm_sync
 from tqdm.asyncio import tqdm as tqdm_async
 
 from mgnipy._models.config import MGnipyConfig
 from mgnipy._shared_helpers.async_helpers import get_semaphore
 from mgnipy.V2.core import MGnifier
+from mgnipy.V2.mixins import StreamMixin
 
 semaphore = get_semaphore()
 BASE_URL = MGnipyConfig().base_url
@@ -36,7 +29,7 @@ BASE_URL = MGnipyConfig().base_url
 # - MGnifyAnalysisDownloadFile OR MGnifyStudyDownloadFile
 
 
-class MGazine:
+class MGazine(StreamMixin):
     """
     Helper for handling MGnify analysis/study downloads.
 
@@ -629,688 +622,6 @@ class MGazine:
             raise ValueError("Either `alias` or `url` must be provided.")
 
         return alias, url
-
-    def stream_tsv(
-        self,
-        url: str,
-        sep: str = "\t",
-        chunksize: Optional[int] = None,
-        max_skip: int = 5,
-        **pd_kwargs,
-    ) -> pd.DataFrame | pd.io.parsers.readers.TextFileReader:
-        """
-        Reads a tsv file from a url and returns an iterator of pandas dataframes.
-        Handles potential issues with extra header rows (causing pd.errors.ParserError)
-        by trying to read the file with increasing skiprows until it succeeds or reaches max_skip.
-
-        Parameters
-        ----------
-        url : str
-            The url of the tsv file to stream.
-        sep : str, optional
-            The separator used in the tsv file. Default is tab.
-        chunksize : int, optional
-            The number of rows to include in each chunk. Default is None.
-        max_skip : int, optional
-            The maximum number of rows to skip before raising an error. Default is 5.
-        pd_kwargs : dict, optional
-            Additional keyword arguments to pass to pandas read_csv.
-
-        Returns
-        -------
-        pd.DataFrame | pd.io.parsers.readers.TextFileReader
-            An iterator of pandas dataframes.
-        """
-
-        """Read a TSV from a URL with resilient header handling.
-
-        This helper will attempt increasing ``skiprows`` when pandas raises a
-        ``ParserError`` (useful for files with extra header lines). The method
-        returns a dataframe or an iterator (when ``chunksize`` is set).
-
-        Notes
-        -----
-        The function returns the same values as ``pandas.read_csv`` and is
-        intentionally conservative in network usage; doctest examples below
-        only exercise the local behaviour when given a filepath-like URL.
-
-        Examples
-        --------
-        >>> # Using a simple local file path (works as a doctest example)
-        >>> import tempfile, pathlib
-        >>> tmp = tempfile.NamedTemporaryFile(mode='w+', delete=False)
-        >>> tmp.write('a\tb\n1\t2\n')
-        >>> tmp.flush()
-        >>> mg = MGazine([{"alias":"x","url":tmp.name,"file_type":"tsv"}])
-        >>> df = mg.stream_tsv(tmp.name)
-        >>> isinstance(next(df) if hasattr(df, '__iter__') and not isinstance(df, pd.DataFrame) else df, (pd.DataFrame,))
-        True
-        """
-
-        for skip in range(max_skip + 1):
-            try:
-                return pd.read_csv(
-                    url,
-                    sep=sep,
-                    chunksize=chunksize,
-                    skiprows=skip if skip > 0 else None,
-                    **pd_kwargs,
-                )
-            except pd.errors.ParserError:
-                continue  # Try next skiprows value
-            except Exception as err:
-                raise RuntimeError(
-                    f"Error reading TSV from {url} with skiprows={skip}"
-                ) from err
-        raise pd.errors.ParserError(
-            f"Failed to parse {url} after skipping up to {max_skip} rows."
-        )
-
-    def stream_html(self, url: str, **web_kwargs) -> bool:
-        """
-        Streams an html file from a url and opens it in the default web browser.
-
-        Parameters
-        ----------
-        url : str
-            The url of the html file to stream.
-        web_kwargs : dict, optional
-            Additional keyword arguments to pass to webbrowser.open.
-
-        Returns
-        -------
-        bool
-            True if the url was opened successfully, False otherwise.
-        """
-        """Open an HTML URL in the default web browser.
-
-        Examples
-        --------
-        >>> # The doctest below only demonstrates the callable returns a bool
-        >>> MGazine([{"alias":"x","url":"http://example.com", "file_type":"html"}]).stream_html('http://example.com') in (True, False)
-        True
-        """
-
-        return webbrowser.open(url, **web_kwargs)
-
-    def stream_txt(
-        self,
-        url: str,
-        chunksize: Optional[int] = None,
-        httpx_client: Optional[httpx.Client] = None,
-        **httpx_kwargs,
-    ) -> Generator:
-        """
-        Streams a txt file from a url and returns an iterator of strings.
-
-        Parameters
-        ----------
-        url : str
-            The url of the txt file to stream.
-        chunksize : Optional[int], optional
-            The number of characters to include in each chunk. Default is None.
-        httpx_kwargs : dict, optional
-            Additional keyword arguments to pass to the httpx client.
-
-        Returns
-        -------
-        Generator[str, None, None]
-            An iterator of strings.
-        """
-
-        """Stream a plain-text resource.
-
-        When ``chunksize`` is ``None`` the full text is returned as a string.
-        When ``chunksize`` is an integer the function yields lists of lines.
-
-        Examples
-        --------
-        >>> downloads = [{"alias":"x","url":"/dev/null","file_type":"txt"}]
-        >>> mg = MGazine(downloads)
-        >>> isinstance(mg.stream_txt('/dev/null'), str)
-        True
-        """
-
-        client = httpx_client or self._mgnifier_helper().exec.httpx_client
-
-        if chunksize is None:
-            # load as whole
-            with client.get(url, **httpx_kwargs) as response:
-                response.raise_for_status()
-                return response.text
-        elif isinstance(chunksize, int) and chunksize > 0:
-            # load in chunks
-            with client.stream("GET", url, **httpx_kwargs) as response:
-                response.raise_for_status()
-                chunk = []
-                for line in response.iter_text():
-                    chunk.append(line)
-                    if len(chunk) == chunksize:
-                        yield chunk
-                        chunk = []
-                if chunk:
-                    yield chunk
-        else:
-            raise ValueError("`chunksize` must be a positive integer or None.")
-
-    def stream_fasta(self, url: str, **skbio_kwargs) -> Generator:
-        """
-        Streams a fasta file from a url and returns an iterator of tuples (header, sequence).
-
-        Parameters
-        ----------
-        url : str
-            The url of the fasta file to stream.
-        skbio_kwargs : dict, optional
-            Additional keyword arguments to pass to the skbio parsers.
-            https://scikit.bio/docs/latest/generated/skbio.io.format.fasta.html
-
-        Returns
-        -------
-        Generator[tuple[str, str], None, None]
-            An iterator of tuples (header, sequence).
-        """
-        return read(url, format="fasta", **skbio_kwargs)
-
-    def stream_gff(self, url: str, **skbio_kwargs) -> Generator:
-        """
-        Streams a gff file from a url and returns an iterator of parsed gff records.
-
-        Parameters
-        ----------
-        url : str
-            The url of the gff file to stream.
-        skbio_kwargs : dict, optional
-            Additional keyword arguments to pass to the skbio parser.
-            https://scikit.bio/docs/latest/generated/skbio.io.format.gff3.html
-
-        Returns
-        -------
-        Generator[skbio.io._gff3.GFF3Record, None, None]
-            "generator of tuple (seq_id of str type, skbio.metadata.IntervalMetadata)"
-        """
-        return read(url, format="gff3", **skbio_kwargs)
-
-    def stream_biom(self, url: str, **skbio_kwargs) -> Generator:
-        """
-        Streams a biom file from a url and returns an iterator of parsed biom records.
-
-        Parameters
-        ----------
-        url : str
-            The url of the biom file to stream.
-        skbio_kwargs : dict, optional
-            Additional keyword arguments to pass to the skbio parser.
-        Returns
-        -------
-        Generator[dict, None, None]
-            An iterator of parsed biom records as dictionaries.
-        """
-        return read(url, format="biom", **skbio_kwargs)
-
-    def stream_gzipped(
-        self,
-        url: str,
-        chunksize: Optional[int] = None,
-        httpx_client: Optional[httpx.Client] = None,
-        decode: bool = False,
-        encoding: str = "utf-8",
-        errors: str = "replace",
-        **httpx_kwargs,
-    ) -> bytes | str | io.BufferedReader | io.TextIOWrapper:
-        """
-        Streams a gzipped file from a url and returns a file-like object that can be read in chunks.
-        Written using GPT-5.3-Codex.
-        Uses httpx for streaming and zlib for decompression.
-
-        Parameters
-        ----------
-        url : str
-            The url of the gzipped file to stream.
-        chunksize : int, optional
-            The size of each chunk to read from the stream.
-        httpx_client : httpx.Client, optional
-            The httpx client to use for streaming.
-        decode : bool, default False
-            Whether to decode the decompressed bytes to a string.
-        encoding : str, default "utf-8"
-            The encoding to use for decoding bytes to a string.
-        errors : str, default "replace"
-            The error handling strategy for decoding bytes to a string.
-        **httpx_kwargs : dict
-            Additional keyword arguments to pass to the httpx client.
-
-        Returns
-        -------
-        bytes | str | io.BufferedReader | io.TextIOWrapper
-            A file-like object that can be read in chunks.
-            If `chunksize` is None, returns the full decompressed content as bytes,
-            or string based on `decode`.
-        """
-
-        # Pick caller-provided client, or fallback to the shared MGnifier client.
-        client = httpx_client or self._mgnifier_helper().exec.httpx_client
-        logging.debug(
-            f"stream_gzipped called url={url} chunksize={chunksize} decode={decode}"
-        )
-
-        # Backward-compatible mode: no chunksize means full download into memory.
-        if chunksize is None:
-            logging.debug("Using full-download mode (chunksize=None)")
-            # Perform a normal blocking GET.
-            r = client.get(url, timeout=None, **httpx_kwargs)
-            # Raise if HTTP status is not 2xx.
-            r.raise_for_status()
-            # Create gzip-compatible streaming decompressor (16 + MAX_WBITS enables gzip header).
-            decompressor = zlib.decompressobj(16 + zlib.MAX_WBITS)
-            # Decompress full payload bytes and flush remaining tail bytes.
-            data = decompressor.decompress(r.content) + decompressor.flush()
-            logging.debug(
-                f"Full-download mode complete: compressed={len(r.content)} decompressed={len(data)}"
-            )
-            # Return text if decode=True, else raw bytes.
-            return data.decode(encoding, errors=errors) if decode else data
-
-        # Validate chunksize for streaming mode.
-        if not isinstance(chunksize, int) or chunksize <= 0:
-            raise ValueError("`chunksize` must be a positive integer or None.")
-
-        # Custom raw stream that adapts httpx streamed gzip bytes to a readable file-like object.
-        class _HTTPGzipRaw(io.RawIOBase):
-            def __init__(self):
-                # Open streaming HTTP response context.
-                self._cm = client.stream("GET", url, timeout=None, **httpx_kwargs)
-                # Enter context manually so this object controls lifecycle.
-                self._resp = self._cm.__enter__()
-                # Fail fast on non-2xx.
-                self._resp.raise_for_status()
-                # Iterate compressed bytes in fixed-size network chunks.
-                self._iter = self._resp.iter_raw(chunk_size=chunksize)
-                # Incremental gzip decompressor.
-                self._decomp = zlib.decompressobj(16 + zlib.MAX_WBITS)
-                # Internal decompressed byte buffer.
-                self._buf = bytearray()
-                # End-of-stream marker.
-                self._eof = False
-                # Track whether decompressor.flush() has been called.
-                self._flushed = False
-                logging.debug("Streaming HTTP/gzip reader initialized")
-
-            def readable(self) -> bool:
-                # Required by BufferedReader to know this raw stream supports reading.
-                return True
-
-            def _fill(self, need: int) -> None:
-                # Keep buffering until we have enough bytes or we hit EOF.
-                while len(self._buf) < need and not self._eof:
-                    try:
-                        # Pull next compressed network chunk.
-                        chunk = next(self._iter)
-                    except StopIteration:
-                        # Network stream finished; flush remaining decompressor tail once.
-                        if not self._flushed:
-                            tail = self._decomp.flush()
-                            if tail:
-                                self._buf.extend(tail)
-                            self._flushed = True
-                        # Mark EOF so future reads stop pulling.
-                        self._eof = True
-                        logging.debug("Reached end of HTTP stream")
-                        break
-
-                    # If chunk is non-empty, incrementally decompress and append output.
-                    if chunk:
-                        out = self._decomp.decompress(chunk)
-                        if out:
-                            self._buf.extend(out)
-
-            def readinto(self, b) -> int:
-                # If already closed, signal EOF.
-                if self.closed:
-                    return 0
-                # Create writable view into caller-provided buffer.
-                mv = memoryview(b)
-                # Ensure internal buffer has enough data for requested read.
-                self._fill(len(mv))
-                # Compute how many bytes we can actually return.
-                n = min(len(mv), len(self._buf))
-                # No bytes available means EOF.
-                if n <= 0:
-                    return 0
-                # Copy decompressed bytes into caller buffer.
-                mv[:n] = self._buf[:n]
-                # Remove consumed bytes from internal buffer.
-                del self._buf[:n]
-                # Return byte count copied.
-                return n
-
-            def close(self) -> None:
-                # Close HTTP stream context exactly once.
-                if not self.closed:
-                    try:
-                        self._cm.__exit__(None, None, None)
-                    finally:
-                        super().close()
-                        logging.debug("Streaming HTTP/gzip reader closed")
-
-        # Wrap raw stream in BufferedReader for efficient file-like reads.
-        raw = _HTTPGzipRaw()
-        buffered = io.BufferedReader(raw, buffer_size=chunksize)
-
-        # If decode requested, wrap bytes stream as text stream.
-        if decode:
-            return io.TextIOWrapper(buffered, encoding=encoding, errors=errors)
-
-        # Default return: binary buffered reader (works with ijson.kvitems).
-        return buffered
-
-    def stream_jsonl(
-        self,
-        url: str,
-        orient: Optional[
-            Literal["records", "split", "index", "columns", "values", "table"]
-        ] = None,
-        chunksize: Optional[int] = None,
-        **pd_kwargs,
-    ) -> dict:
-        """
-        Streams a jsonl file from a url and returns the parsed json as a dictionary.
-
-        Parameters
-        ----------
-        url : str
-            The url of the json file to stream.
-        sep : str, optional
-            The separator to use when parsing the json file. Default is "\t".
-        chunksize : Optional[int], optional
-            The size of the chunks to read from the stream. Default is None.
-        max_skip : int, optional
-            The maximum number of rows to skip before raising an error. Default is 5.
-        **pd_kwargs : dict
-            Additional keyword arguments to pass to the pandas parser.
-        Returns
-        -------
-        dict
-            The parsed json as a dictionary.
-        """
-
-        return pd.read_json(
-            url, orient=orient, lines=True, chunksize=chunksize, **pd_kwargs
-        )
-
-    def stream_json(
-        self,
-        url: str,
-        chunksize: Optional[int] = None,
-        httpx_client: Optional[httpx.Client] = None,
-        **httpx_kwargs,
-    ) -> dict | Generator:
-        """
-        Streams a json file from a url and returns the parsed json as a dictionary or an iterator of dictionaries if chunksize is specified.
-
-        Parameters
-        ----------
-        url : str
-            The url of the json file to stream.
-        chunksize : Optional[int], optional
-            The size of the chunks to read from the stream. Default is None.
-        **httpx_kwargs : dict
-            Additional keyword arguments to pass to the httpx client.
-        Returns
-        -------
-        dict | Generator
-            The parsed json as a dictionary, or an iterator of dictionaries if chunksize is specified.
-        """
-
-        client = httpx_client or self._mgnifier_helper().exec.httpx_client
-
-        # normal full get and json parse
-        if chunksize is None and not (url.endswith(".gz") or url.endswith(".gzip")):
-            # If no chunksize and not gzipped, do a normal full GET and parse as JSON.
-            with client.get(url, **httpx_kwargs) as response:
-                response.raise_for_status()
-                return response.json()
-        # streaming json, not zipped
-        elif chunksize is not None and not (
-            url.endswith(".gz") or url.endswith(".gzip")
-        ):
-            # If chunksize specified and not gzipped, stream text and parse JSON objects one by one.
-            with client.stream("GET", url, **httpx_kwargs) as response:
-                response.raise_for_status()
-                for entry in ijson.kvitems(response.iter_text(), ""):
-                    yield entry
-        # gzipped json (with or without chunksize)
-        elif url.endswith(".gz") or url.endswith(".gzip"):
-            # If gzipped, use the stream_gzipped method to get a file-like object and parse JSON objects one by one.
-            with self.stream_gzipped(
-                url,
-                chunksize=chunksize,
-                httpx_client=client,
-                decode=True,
-                **httpx_kwargs,
-            ) as gzipped_stream:
-                for entry in ijson.kvitems(gzipped_stream, ""):
-                    yield entry
-        else:
-            raise ValueError(f"Unsupported file type for URL: {url}")
-
-    def stream_tree(self, url: str, **skbio_kwargs) -> Generator:
-        """
-        Streams a tree file from a url and returns an iterator of parsed tree records.
-
-        Parameters
-        ----------
-        url : str
-            The url of the tree file to stream.
-        skbio_kwargs : dict, optional
-            Additional keyword arguments to pass to the skbio parser.
-        Returns
-        -------
-        Generator[dict, None, None]
-            An iterator of parsed tree records as dictionaries.
-        """
-        return read(url, format="newick", **skbio_kwargs)
-
-    def _fix_inconsistent_cols(
-        self, fields: list[str], pad_to: int = 15
-    ) -> list[str] | None:
-        """
-        Fixes inconsistent columns in a list of strings.
-
-        Parameters
-        ----------
-        fields : list[str]
-            The list of strings to fix.
-        pad_to : int, optional
-            The number of columns to pad or truncate to. Default is 15.
-
-        Returns
-        -------
-        list[str] | None
-            The fixed list of strings, or None if the fields are invalid.
-        """
-        """Pad or truncate list of fields to ``pad_to`` length.
-
-        Examples
-        --------
-        >>> MGazine([])._fix_inconsistent_cols(['a','b'], pad_to=4)
-        ['a', 'b', '', '']
-        >>> MGazine([])._fix_inconsistent_cols(list('ABCDEFGHI'), pad_to=3)
-        ['A', 'B', 'C']
-        """
-
-        # pad to
-        if len(fields) < pad_to:
-            return fields + [""] * (pad_to - len(fields))
-        # truncate
-        if len(fields) > pad_to:
-            return fields[:pad_to]
-        return fields
-
-    def _get_streamer(
-        self,
-        alias: Optional[str] = None,
-        url: Optional[HttpUrl] = None,
-        chunksize: int = 1000,
-        httpx_client: Optional[httpx.Client] = None,
-        max_skip: int = 5,
-        **kwargs,
-    ):
-        """
-        Gets the appropriate streamer function based on the file type of the download.
-
-        Parameters
-        ----------
-        alias : Optional[str]
-            The alias of the download to stream.
-        url : Optional[HttpUrl]
-            The url of the download to stream.
-        chunksize : int, optional
-            The size of the chunks to read from the stream. Default is 1000.
-        httpx_client : Optional[httpx.Client], optional
-            The httpx client to use for making requests. Default is None.
-        max_skip : int, optional
-            The maximum number of rows to skip before raising an error. Default is 5.
-        kwargs : dict
-            Additional keyword arguments to pass to the streamer function.
-
-        Returns
-        -------
-        Callable
-            A function that can be called to stream the download.
-        """
-
-        client = httpx_client or self._mgnifier_helper().exec.httpx_client
-
-        # get alias/url
-        _alias, _url = self._prioritize_alias(alias, url, required=True)
-        # return stream based on file type
-        file_type = self._get_type_by_alias(_alias)
-        # if a tsv, check if gzipped first, then try normal tsv, with error handling for inconsistent columns
-        if file_type == "tsv":
-            if _url.endswith(".gz") or _url.endswith(".gzip"):
-                logging.debug(f"tsv file type ends with .gz: {_url}")
-                try:
-                    return self.stream_tsv(
-                        _url,
-                        chunksize=chunksize,
-                        max_skip=max_skip,
-                        compression="gzip",
-                        **kwargs,
-                    )
-                except pd.errors.ParserError as e:
-                    logging.error(f"ParserError: {e}")
-                    return self.stream_tsv(
-                        _url,
-                        chunksize=chunksize,
-                        max_skip=max_skip,
-                        compression="gzip",
-                        engine="python",
-                        on_bad_lines=self._fix_inconsistent_cols,
-                        **kwargs,
-                    )
-            elif _url.endswith(".txt") or _url.endswith(".tsv"):
-                return self.stream_tsv(
-                    _url, chunksize=chunksize, max_skip=max_skip, **kwargs
-                )
-        # if a csv, try normal csv, with error handling for inconsistent columns
-        elif file_type == "csv":
-            return self.stream_tsv(
-                _url, sep=",", chunksize=chunksize, max_skip=max_skip, **kwargs
-            )
-        # if html, open in web browser (not really streaming but for consistency with the interface)
-        elif file_type == "html":
-            return lambda: self.stream_html(_url, **kwargs)
-        # if txt
-        elif file_type == "txt":  # TODO: to constants
-            return self.stream_txt(
-                _url, chunksize=chunksize, httpx_client=client, **kwargs
-            )
-        elif file_type == "gff":
-            return self.stream_gff(_url, **kwargs)
-        elif file_type == "biom":
-            return self.stream_biom(_url, **kwargs)
-        elif file_type == "fasta":
-            return self.stream_fasta(_url, **kwargs)
-        elif file_type == "tree":
-            return self.stream_tree(_url, **kwargs)
-        elif file_type == "json":
-            return self.stream_jsonl(
-                _url, orient="records", chunksize=chunksize, **kwargs
-            )
-        elif file_type == "other" and ".json" in _url:
-            if _url.endswith("json.gz") or _url.endswith("json.gzip"):
-                return self.stream_json(
-                    _url, chunksize=chunksize, httpx_client=client, **kwargs
-                )
-
-            logging.info(
-                f"{_alias} is only available for download "
-                f"(e.g., `.download({_alias}))`"
-            )
-            # more info
-            logging.debug(
-                f"Alias: {_alias}\nURL: {_url}\nFile type: {file_type}. "
-                "Only '.json' files can be streamed under 'other' type, "
-                "otherwise this download is only available for download."
-            )
-
-        else:
-            raise ValueError(f"Unsupported file type for streaming: {file_type}")
-
-    def stream(
-        self,
-        *,
-        alias: Optional[str] = None,
-        url: Optional[HttpUrl] = None,
-        chunksize: Optional[int] = None,
-        max_skip: int = 5,
-        **kwargs,
-    ) -> Any:
-        """
-        Streams a single download based on its alias or url.
-
-        If ``chunksize`` is specified then iterators of dataframes or strings
-        will be returned; otherwise the full data will be returned as a single
-        object.
-
-        Parameters
-        ----------
-        alias : Optional[str]
-            The alias of the download to stream.
-        url : Optional[HttpUrl]
-            The url of the download to stream.
-        chunksize : Optional[int]
-            The size of the chunks to read from the stream.
-        max_skip : int, optional
-            The maximum number of rows to skip before raising an error. Default is 5.
-        **kwargs
-            Additional keyword arguments to pass to the streamer function.
-
-        Returns
-        -------
-        Any
-            The streamer result for the resolved alias or url.
-        """
-        # resolve a single alias/url target
-        _alias, _url = self._prioritize_alias(alias, url, required=True)
-
-        # return a single streamer result, not a dict of all streams
-        client = self._mgnifier_helper().exec.httpx_client
-        logging.info("Setting up stream for alias=%s url=%s", _alias, _url)
-
-        try:
-            return self._get_streamer(
-                alias=_alias,
-                url=_url,
-                chunksize=chunksize,
-                httpx_client=client,
-                max_skip=max_skip,
-                **kwargs,
-            )
-        except HTTPError as err:
-            logging.error("HTTP error for alias=%s url=%s: %s", _alias, _url, err)
-            raise
 
 
 class MGazineCurator:
